+++ CHANGELOG.md
@@ -13,6 +13,7 @@
 - [Feature] Add dialog about sucs/not update
 - [Feature] Add failed upload and network dialog
 - [Feature] Add synchronization percent in archive screen
+- [Feature] Change percent scheme for synchronization
 - [BUGFIX] Fix incorrect corner radius for battery icon
 - [BUGFIX] Fix searching pic
 - [CI] Pass countly prod creds in application when building